+++ thyrel-api/thyrel-api/Websocket/WebsocketHandler.cs
@@ -7,6 +7,7 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using thyrel_api.Controllers;
+using thyrel_api.DataProvider;
 using thyrel_api.Models;
 
 namespace thyrel_api.Websocket
